The moderately great numAddrs removal:
numAddrs is now transient instead of persistent, and there is a new decodeInode
function to be used instead of open Data.Serialize.decode.

This was done because persisting the numAddrs field only for some sanity checks
and simple math in the Serialize (Inode t) instance was a bit yucky.

Tests modified accordingly.