BUG checkpoint: propM_basicWRWR raises the bug with 3 writes of particular size; neither dirMutexOK nor sync have anything to do with the problem.
Details:

propM_dirMutexOK was triggering an assert in the guts of Inode.allocFill, but
rarely.  After obtaining the failing test case (the important data point of
which was the lengths of the data being written, originating from writing
randomly-created directory names for each of 8 threads), I made propM_dirMutexOK
test sequential to be sure there wasn't a race condition being triggered, which
there wasn't.

More logging revealed the bad sequence of writes happening, which was then
narrowed considerably, from many hundreds of writes to just the writes currently
reflected in propM_basicWRWR.