Replace "Cont" with "Ext" throughout the code base.
We unfortunately used a name "Cont" as a core data structure -- the idea was
that it was to represent follow-on data to be associated with a particular inode
"chain" of these fixed-size "Conts".  It's essentially just a singly-linked
list, but the names "Cont" and "continuation" are highly misleading for obvious
reasons.

We briefly considered "Link" instead of "Cont" but that'd also be awkwardly
overloaded: filesystem hard/soft links are going to be named as such elsewhere.

Finally, feeling rather uninspired, we decided to call this datatype simply an
"Ext" -- or an "inode extension" if you will.  Thus we can now talk about
ExtRefs, Exts, and nextExt links inside inodes, and the like.  Perhaps not the
BEST naming convention possible, but it was just too hard to conscion the
confusing and pervasive use of "Cont".